TailSHINE_BACKEND/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
currency-converter-with-rate==1.1.0
cycler==0.11.0
deep-translator==1.8.3
Django==4.0.4
django-cors-headers==3.11.0
django-crontab==0.7.1
django-extensions==3.1.5
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==4.8.0
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.1
h11==0.9.0
h2==3.2.0
halo==0.0.31
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.19.0
itypes==1.2.0
Jinja2==3.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langdetect==1.0.9
language-tool-python==2.7.1
libclang==14.0.1
log-symbols==0.0.14
Markdown==3.3.7
MarkupSafe==2.1.1
multidict==6.0.2
networkx==2.8
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.8
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-whois==0.7.3
pytz==2022.1
PyWavelets==1.3.0
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spinners==0.0.24
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2022.5.4
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
xmltodict==0.12.0
yarl==1.7.2
